added cadasta dialog parent class
