+++ cadasta/gui/tools/cadasta_login.py
@@ -1,45 +1,28 @@
 # -*- coding: utf-8 -*-
-"""
-/***************************************************************************
- CadastaDialog
-                                 A QGIS plugin
- This tool helps create, update, upload and download Cadasta projects.
-                             -------------------
-        begin                : 2016-11-25
-        git sha              : $Format:%H$
-        copyright            : (C) 2016 by Kartoza
-        email                : christian@kartoza.com
- ***************************************************************************/
+"""Contains login dialog
 
-/***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
+.. note:: This program is free software; you can redistribute it and/or modify
+     it under the terms of the GNU General Public License as published by
+     the Free Software Foundation; either version 2 of the License, or
+     (at your option) any later version.
+
 """
 
-import logging
-from qgis.PyQt import QtGui
 from qgis.gui import QgsMessageBar
 from cadasta.utilities.resources import get_ui_class
 
+from cadasta.gui.tools.cadasta_dialog import CadastaDialog
 from cadasta.api.login import Login
-from cadasta.gui.tools.cadasta_style import CadastaStyle
 from cadasta.common.setting import save_authtoken, save_url_instance
 
-LOGGER = logging.getLogger('CadastaQGISPlugin')
 FORM_CLASS = get_ui_class('cadasta_login_base.ui')
 
 
-class CadastaLogin(QtGui.QDialog, FORM_CLASS):
-    def __init__(self, parent=None):
+class CadastaLogin(CadastaDialog, FORM_CLASS):
+    def __init__(self):
         """ Constructor."""
 
-        super(CadastaLogin, self).__init__(parent)
-        self.setupUi(self)
+        super(CadastaLogin, self).__init__()
         self.message_bar = None
         self.text_test_connection_button = self.test_connection_button.text()
         self.ok_label.setVisible(False)
@@ -47,36 +30,10 @@ class CadastaLogin(QtGui.QDialog, FORM_CLASS):
 
     def init_style(self):
         """ Initiate custom styles for dialog. """
-
-        self.setStyleSheet('background-color:white')
         self.disable_button(self.save_button)
         self.enable_button(self.test_connection_button)
         self.test_connection_button.clicked.connect(self.login)
         self.save_button.clicked.connect(self.save_authtoken)
-
-    def enable_button(self, custom_button):
-        """ Enable button.
-
-        :param custom_button: button that is enabled
-        :type custom_button: QWidget
-        """
-
-        custom_button.setEnabled(True)
-        custom_button.setStyleSheet(
-                'background-color:#525252; cursor:pointer;' +
-                CadastaStyle.button_style()
-        )
-
-    def disable_button(self, custom_button):
-        """ Disable button.
-
-        :param custom_button: button that is enabled
-        :type custom_button: QWidget
-        """
-        custom_button.setStyleSheet(
-                'background-color:#A8A8A8;' + CadastaStyle.button_style()
-        )
-        custom_button.setEnabled(False)
 
     def login(self):
         """ Login function when tools button clicked."""
@@ -92,18 +49,18 @@ class CadastaLogin(QtGui.QDialog, FORM_CLASS):
         if not self.url or not username or not password:
             self.message_bar = QgsMessageBar()
             self.message_bar.pushWarning(
-                    self.tr('Error'),
-                    self.tr('URL/Username/password is empty.')
+                self.tr('Error'),
+                self.tr('URL/Username/password is empty.')
             )
         else:
             self.disable_button(self.test_connection_button)
             self.test_connection_button.setText(self.tr('Logging in...'))
             # call tools API
             self.login_api = Login(
-                    self.url,
-                    username,
-                    password,
-                    self.on_finished)
+                self.url,
+                username,
+                password,
+                self.on_finished)
 
     def on_finished(self, result):
         """ On finished function when tools request is finished."""
